Refactored with base tI dashboard class
