+++ browser-test/src/trusted_intermediary.test.ts
@@ -979,38 +979,46 @@ test.describe('Trusted intermediaries', {tag: ['@uses-fixtures']}, () => {
       page,
       tiDashboard,
     }) => {
-      await loginAsTrustedIntermediary(page)
-      await tiDashboard.gotoTIDashboardPage(page)
-      await waitForPageJsLoad(page)
+      await test.step('Login as a TI and go to the dashboard', async () => {
+        await loginAsTrustedIntermediary(page)
+        await tiDashboard.gotoTIDashboardPage(page)
+        await waitForPageJsLoad(page)
+      })
 
-      await validateScreenshot(
-        page.getByTestId('ti-nav'),
-        'ti-navigation-dashboard',
-      )
+      await test.step('Validate that TI navigation renders correctly with aria-current on dashboard page', async () => {
+        await validateScreenshot(
+          page.getByTestId('ti-nav'),
+          'ti-navigation-dashboard',
+        )
 
-      // When we're on the client list page (the dashboard), the current tab should be the client list tab.
-      expect(
-        await page.getAttribute('#account-settings-link', 'aria-current'),
-      ).toBeNull()
-      expect(
-        await page.getAttribute('#client-list-link', 'aria-current'),
-      ).not.toBeNull()
+        // When we're on the client list page (the dashboard), the current tab should be the client list tab.
+        expect(
+          await page.getAttribute('#account-settings-link', 'aria-current'),
+        ).toBeNull()
+        expect(
+          await page.getAttribute('#client-list-link', 'aria-current'),
+        ).not.toBeNull()
+      })
 
-      await tiDashboard.goToAccountSettingsPage(page)
-      await waitForPageJsLoad(page)
+      await test.step('Go to the TI Account Settings page', async () => {
+        await tiDashboard.goToAccountSettingsPage(page)
+        await waitForPageJsLoad(page)
+      })
 
-      await validateScreenshot(
-        page.getByTestId('ti-nav'),
-        'ti-navigation-account-settings',
-      )
+      await test.step('Validate that TI navigation renders correctly with aria-current on account settings page', async () => {
+        await validateScreenshot(
+          page.getByTestId('ti-nav'),
+          'ti-navigation-account-settings',
+        )
 
-      // When we're on the account settings page, the current tab should be the account settings tab.
-      expect(
-        await page.getAttribute('#account-settings-link', 'aria-current'),
-      ).not.toBeNull()
-      expect(
-        await page.getAttribute('#client-list-link', 'aria-current'),
-      ).toBeNull()
+        // When we're on the account settings page, the current tab should be the account settings tab.
+        expect(
+          await page.getAttribute('#account-settings-link', 'aria-current'),
+        ).not.toBeNull()
+        expect(
+          await page.getAttribute('#client-list-link', 'aria-current'),
+        ).toBeNull()
+      })
     })
   })
